fix: 修复是否为父目录判断错误
