+++ apps/ai/src/server/trpc/route/jobs/apps.ts
@@ -36,12 +36,12 @@ import { procedure } from "src/server/trpc/procedure/base";
 import { checkAppExist, checkCreateAppEntity, getClusterAppConfigs } from "src/server/utils/app";
 import { getAdapterClient } from "src/server/utils/clusters";
 import { clusterNotFound } from "src/server/utils/errors";
-import { isParentOrSameFolder } from "src/server/utils/file";
 import { forkEntityManager } from "src/server/utils/getOrm";
 import { commitContainerImage, createHarborImageUrl, pushImageToHarbor } from "src/server/utils/image";
 import { logger } from "src/server/utils/logger";
 import { paginate, paginationSchema } from "src/server/utils/pagination";
 import { getClusterLoginNode, sshConnect } from "src/server/utils/ssh";
+import { isParentOrSameFolder } from "src/utils/file";
 import { isPortReachable } from "src/utils/isPortReachable";
 import { BASE_PATH } from "src/utils/processEnv";
 import { z } from "zod";
@@ -331,7 +331,9 @@ export const createAppSession = procedure
       const homeDir = await getUserHomedir(ssh, userId, logger);
 
       // 工作目录和挂载点必须在用户的homeDir下
-      if (!isParentOrSameFolder(homeDir, workingDirectory) || !isParentOrSameFolder(homeDir, mountPoint)) {
+
+      if ((workingDirectory && !isParentOrSameFolder(homeDir, workingDirectory))
+        || (mountPoint && !isParentOrSameFolder(homeDir, mountPoint))) {
         throw new TRPCError({
           code: "BAD_REQUEST",
           message: "workingDirectory and mountPoint should be in homeDir",
