+++ apps/ai/src/server/trpc/route/jobs/jobs.ts
@@ -24,10 +24,10 @@ import { procedure } from "src/server/trpc/procedure/base";
 import { checkCreateAppEntity } from "src/server/utils/app";
 import { getAdapterClient } from "src/server/utils/clusters";
 import { clusterNotFound } from "src/server/utils/errors";
-import { isParentOrSameFolder } from "src/server/utils/file";
 import { forkEntityManager } from "src/server/utils/getOrm";
 import { logger } from "src/server/utils/logger";
 import { getClusterLoginNode, sshConnect } from "src/server/utils/ssh";
+import { isParentOrSameFolder } from "src/utils/file";
 import { z } from "zod";
 
 const SESSION_METADATA_NAME = "session.json";
@@ -107,7 +107,7 @@ procedure
 
         const homeDir = await getUserHomedir(ssh, userId, logger);
 
-        if (!isParentOrSameFolder(homeDir, mountPoint)) {
+        if (mountPoint && !isParentOrSameFolder(homeDir, mountPoint)) {
           throw new TRPCError({
             code: "BAD_REQUEST",
             message: "mountPoint should be in homeDir",
